chore: move form error constant in different folder
