+++ src/core/constants/form.ts
@@ -1,3 +1,4 @@
+// Errors
 export enum FORM_ERRORS_ENUM {
   existingCode = 'existingCode',
   invalidGroupValue = 'invalidGroupValue',
